rename remacs-lib to ng-docfile
