+++ Makefile.in
@@ -1204,7 +1204,7 @@ check-declare:
 
 rustfmt:
 	(cd rust_src && cargo fmt)
-	(cd rust_src/remacs-lib && cargo fmt)
+	(cd rust_src/ng-docfile && cargo fmt)
 	(cd rust_src/crates/lisp_util && cargo fmt)
 	(cd rust_src/crates/lisp_macros && cargo fmt)
