+++ rust_src/ng-docfile/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "remacs-lib"
+name = "ng-docfile"
 version = "0.1.0"
 edition = "2018"
